+++ mlflow/genai/experimental/__init__.py
@@ -5,7 +5,9 @@ This module contains experimental features for MLflow GenAI functionality.
 """
 
 from mlflow.genai.experimental.databricks_trace_archival import enable_databricks_trace_archival
+from mlflow.genai.experimental.databricks_trace_exporter import MlflowV3DeltaSpanExporter
 
 __all__ = [
     "enable_databricks_trace_archival",
+    "MlflowV3DeltaSpanExporter",
 ]
